+++ include/bringauto/fleet_protocol/async_function_execution_definitions/ConvertibleBuffer.hpp
@@ -7,7 +7,7 @@
 
 
 
-namespace bringauto::fleet_protocol::cxx {
+namespace bringauto::fleet_protocol::async_function_execution_definitions {
 
 /**
  * @brief ConvertibleBuffer is a helper class to convert a fleet_protocol 'buffer' struct
