+++ include/bringauto/fleet_protocol/async_function_execution_definitions/ConvertibleBufferReturn.hpp
@@ -8,7 +8,7 @@
 
 
 
-namespace bringauto::fleet_protocol::cxx {
+namespace bringauto::fleet_protocol::async_function_execution_definitions {
 
 /**
  * @brief ConvertibleBufferReturn is a helper class to convert a fleet_protocol 'buffer' struct
